pyproject.toml
[tool.poetry]
name = "hw6"
version = "0.1.0"
description = ""
authors = ["Anastasiia2001148 <164914414+Anastasiia2001148@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
faker = "^28.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
